requirements.txt
groq
pandas
openpyxl
pypdf
tqdm
python-dotenv
PyYAML